+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "2.styleguide",
+  "name": "v13-geckos-team-01",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
