+++ src/App.js
@@ -1,24 +1,22 @@
 import React from 'react';
-import logo from './logo.svg';
-import './App.css';
+import { BrowserRouter as Router, Route, Switch } from "react-router-dom";
+
+import Sample from './pages/sample'
+import GetStarted from './pages/getStarted'
+import AddItUp from './pages/addItUp'
+import Results from './pages/results'
 
 function App() {
   return (
     <div className="App">
-      <header className="App-header">
-        <img src={logo} className="App-logo" alt="logo" />
-        <p>
-          Edit <code>src/App.js</code> and save to reload.
-        </p>
-        <a
-          className="App-link"
-          href="https://reactjs.org"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          Learn React
-        </a>
-      </header>
+          <Router>
+            <Switch>
+              <Route exact path='/' component={ GetStarted }/>
+              <Route exact path='/search' component={ AddItUp }/>
+              <Route exact path='/results' component={ Results }/>
+              <Route exact path='/sample' component={ Sample }/>
+              </Switch>
+          </Router>
     </div>
   );
 }
