--- src/Components/SampleWebsite/MainNavbar.jsx
@@ -1,35 +1,0 @@
-import React from 'react'
-
-import { Navbar, Nav, NavDropdown, Dropdown } from 'react-bootstrap'
-
-
-
-import './styles.css'
-
-
-function MainNavbar({history}) {
-    return (
-        <div>
-        <Navbar expand="lg" className="color3">
-            <Navbar.Brand href="#home">LogoHere</Navbar.Brand>
-            <Navbar.Toggle aria-controls="basic-navbar-nav" />
-            <Navbar.Collapse id="basic-navbar-nav" >
-                <Nav className="mr-auto">
-                    <Nav.Link href="/" className="letterColorNavbar">Home</Nav.Link>
-                    <NavDropdown className="letterColorNavbar" title="Dropdown">
-                    <NavDropdown.Item href="#action/3.1">Brand Essence</NavDropdown.Item>
-                    <NavDropdown.Item href="#action/3.2">Proposed Logo</NavDropdown.Item>
-                    <NavDropdown.Item href="#action/3.3">Proposed Color palette</NavDropdown.Item>
-                    <NavDropdown.Item href="#action/3.3">Proposed Fonts</NavDropdown.Item>
-                    <NavDropdown.Item href="#action/3.3">Proposed Icons</NavDropdown.Item>
-                    <NavDropdown.Item href="#action/3.3">Proposed Forms and buttons</NavDropdown.Item>
-                </NavDropdown>
-                </Nav>
-             
-            </Navbar.Collapse>
-</Navbar>
-</div>
-    )
-}
-
-export default MainNavbar;
