--- src/Components/SampleWebsite/Sample.jsx
@@ -1,15 +1,0 @@
-import React from 'react'
-
-import MainNavbar from './MainNavbar';
-import Title from './Title';
-
-function Sample() {
-    return (
-        <div>
-            <MainNavbar/>
-            <Title/>
-        </div>
-    )
-}
-
-export default Sample;
