+++ src/index.js
@@ -1,3 +1,6 @@
+//import react.bootstrap stylesheet
+import 'bootstrap/dist/css/bootstrap.min.css';
+
 //import react imports
 import React from 'react';
 import ReactDOM from 'react-dom';
@@ -5,16 +8,5 @@ import ReactDOM from 'react-dom';
 //import children App component
 import App from './App';
 
-import { BrowserRouter } from "react-router-dom";
-
-//import react.bootstrap stylesheet
-import 'bootstrap/dist/css/bootstrap.min.css';
-
-
-ReactDOM.render(
-
-<BrowserRouter>
-<App />
-</BrowserRouter>
-, document.getElementById('root'));
+ReactDOM.render(<App />, document.getElementById('root'));
