+++ src/pages/addItUp.jsx
@@ -1,0 +1,22 @@
+import React from 'react';
+
+import { Container } from 'react-bootstrap'
+
+//import components
+import MainNavbar from '../components/MainNavbar'
+import Title from '../components/Title'
+import Brand from '../components/Brand'
+import Footer from '../components/Footer'
+
+function AddItUp() {
+  return (
+      <Container fluid="true">
+          <MainNavbar/>
+          <Title/>
+          <Brand/>
+          <Footer/>
+      </Container>
+  );
+}
+
+export default AddItUp;
