+++ src/pages/sample.jsx
@@ -2,24 +2,22 @@ import React from 'react';
 
 import { Container } from 'react-bootstrap'
 
-
 //import components
-import MainNavbar from './MainNavbar'
-import Title from './Title'
-import Brand from './Brand'
-import Logos from './Logos'
-import Colors from './Colors'
-import Typography from './Typography'
-import Icons from './Icons'
-import MainForm from './Form'
-import Footer from './Footer'
+import MainNavbar from '../components/MainNavbar'
+import Title from '../components/Title'
+import Brand from '../components/Brand'
+import Logos from '../components/Sample/Logos'
+import Colors from '../components/Sample/Colors'
+import Typography from '../components/Sample/Typography'
+import Icons from '../components/Sample/Icons'
+import MainForm from '../components/Sample/Form'
+import Footer from '../components/Footer'
 
 function ComponentIndex() {
   return (
       <Container fluid="true">
           <MainNavbar/>
           <Title/>
-          <br/>
           <Brand/>
           <br/>
           <Logos/>
